+++ src/page/ErrorPage.js
@@ -1,5 +1,5 @@
 import { render } from "../main";
-import { createNodeElement, renderByNodeElement } from "../utils";
+import { createNodeElement, renderByNodeElement } from "./utils";
 
 export default () => {
   const component = `
