+++ src/page/MainPage.js
@@ -1,7 +1,7 @@
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import { render } from "../main";
-import { createNodeElement, renderByNodeElement } from "../utils";
+import { createNodeElement, renderByNodeElement } from "./utils";
 import auth from "../auth";
 
 export default () => {
