+++ src/page/ProfilePage.js
@@ -1,8 +1,8 @@
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import auth from "../auth";
 import { render } from "../main";
-import { createNodeElement, renderByNodeElement } from "../utils";
+import { createNodeElement, renderByNodeElement } from "./utils";
 
 export default () => {
   const { username, email, bio } = auth.getUser();
